Create Seprate Controller for Product work + Add cut price in product + some links changing
