+++ resources/views/web/pages/home.blade.php
@@ -482,6 +482,7 @@
                                         <h2 class="product-title"><a href="{{ route('web.product', ['id' => $val['id']]) }}">{{ $val['title'] }}</a></h2>
                                         <div class="product-price">
                                             <span>£{{ $val['price'] }}</span>
+                                            <del>{{ $val['cut_price'] ? '£'.$val['cut_price'] : NULL}}</del>
                                         </div>
                                     </div>
                                 </div>
