+++ resources/views/web/pages/shop.blade.php
@@ -113,7 +113,7 @@
                                                 <h2 class="product-title"><a href="{{ route('web.product', ['id' => $val->id]) }}">{{ $val->title }}</a></h2>
                                                 <div class="product-price">
                                                     <span>£{{ $val->price }}</span>
-                                                    {{-- <del>$162.00</del> --}}
+                                                    <del>{{ $val->cut_price ? '£'.$val->cut_price : NULL }}</del>
                                                 </div>
                                             </div>
                                         </div>
